Add keyword to AutoReg() to avoid Future Warning
In AutoReg().fit() call in train_lv() add keyword old_names=False to stop
Future Warning from appearing every time the function is called. The
Future Warning is part of statsmodels v 0.12.2 because parameter
names will change in the newer versions. The AR parameter estimates are
not affected by the introduction of this keyword.