Add keyword to AutoReg() to avoid Future Warning (#20)
* Add keyword to AutoReg() to avoid Future Warning

In AutoReg().fit() call in train_lv() add keyword old_names=False to stop
Future Warning from appearing every time the function is called. The
Future Warning is part of statsmodels v 0.12.2 because parameter
names will change in the newer versions. The AR parameter estimates are
not affected by the introduction of this keyword.

* remove explanation for old_names keyword

As @znicholls pointed out, an explanation comment is not necessary because the explanation is documented in the PR anyways (plus can just look it up in case I update package and get different warning)