Change approach for fitting AR parameters
Instead of pooling all runs and estimating the AR order and params from
a single time series, the AR order and params are estimated in a 2 step
approach in which we always fit on individual runs (which makes more
sense from a physical perspective and avoids statistical artefacts)

First the AR order is determined based on the BIC criterion on each run
of a specific scenario. The median estimated order is the order taken
for this scenario. Once the order for each scenario has been determined
the overall median order is selected for the global variability module
of the ESM at hand.

For the selected order AR parameters are determined for each run of each
scenario individually. First the average across all runs within 1
scenario is taken and then the average across all scenarios.

This approach thus assigns equal weight to all scenarios. In the future
the ability to weigh samples differently could and should be added.

Finally a check on whether the estimated paramters constitute a
stationary AR process is carried out.