Update syntax and add AR(0) drawing possibility
One of the keys in params_gv changed its name and content as part of the
train_gv() update. This is now changed here too.

The ability to draw from an AR(0) process is added (also because chances
are now higher that an AR(0) process could be selected than they used to
be before when all runs were pooled)