Improve AR processes aspects (#76)
* Change approach for fitting AR parameters

Instead of pooling all runs and estimating the AR order and params from
a single time series, the AR order and params are estimated in a 2 step
approach in which we always fit on individual runs (which makes more
sense from a physical perspective and avoids statistical artefacts)

First the AR order is determined based on the BIC criterion on each run
of a specific scenario. The median estimated order is the order taken
for this scenario. Once the order for each scenario has been determined
the overall median order is selected for the global variability module
of the ESM at hand.

For the selected order AR parameters are determined for each run of each
scenario individually. First the average across all runs within 1
scenario is taken and then the average across all scenarios.

This approach thus assigns equal weight to all scenarios. In the future
the ability to weigh samples differently could and should be added.

Finally a check on whether the estimated paramters constitute a
stationary AR process is carried out.

* Update syntax and add AR(0) drawing possibility

One of the keys in params_gv changed its name and content as part of the
train_gv() update. This is now changed here too.

The ability to draw from an AR(0) process is added (also because chances
are now higher that an AR(0) process could be selected than they used to
be before when all runs were pooled)

* fix docstring

* fix docstring

* fix line lengths

* Fix a comment

* Yet another tiny docstring change

* Add check for equal weight and improve formatting

Add TODO to docstring about accepting different weighting of samples

Check that each scen is supposed to be weighted equally, else raise
error

Introduce correct line length in docstring and comments

* Update test output files

Update the output files of test_calibrate_mesmer() and
test_draw_realisations_from_bundle() which both changed because of the
improved AR parameter estimation approach in train_gv().

* Turn max_lag and sel_crit into keyword argument

Based on suggestion from @znicholls

* Fix typo in comment of create_emus_gv()

Co-authored-by: Zeb Nicholls <zebedee.nicholls@climate-energy-college.org>

* Update kwargs entry in docstring train_gv()

Co-authored-by: Zeb Nicholls <zebedee.nicholls@climate-energy-college.org>

Co-authored-by: Zeb Nicholls <zebedee.nicholls@climate-energy-college.org>