Remove faulty EC-EARTH run
r14 of rcp45 of EC-EARTH has a faulty time axis in the cmip5-ng archive
with 10 years always sharing the same time stamp. Bug is reported which
will likely eventually lead to other changes (i.e., this exclusion will
become irrelevant again because either the run will be removed from the
archive or it will be fixed).