Salvage cmip5-ng (#90)
* Fix time axis issue when reading in cmip5-ng data

The time encoding in cmip5-ng is apparently not compatible with the
default approach used by xarray 0.17.0 to read in data. Hence, the use_cftime
argument needed to be introduced.

* Add ability to have no r1i1 run

In case no *r1i1* run exists leave the runs in their original order.
Only reason *r1i1* is moved to front is to make it possible to compute
anomalies by baselining every run with respect to the first run (i.e.,
a legacy from the original MESMER description paper approach of training
on a single run).

* Remove faulty EC-EARTH run

r14 of rcp45 of EC-EARTH has a faulty time axis in the cmip5-ng archive
with 10 years always sharing the same time stamp. Bug is reported which
will likely eventually lead to other changes (i.e., this exclusion will
become irrelevant again because either the run will be removed from the
archive or it will be fixed).

* Update CHANGELOG.rst